% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_giedata2input}
\alias{check_giedata2input}
\title{check_giedata2input}
\usage{
check_giedata2input(
  countries,
  companies,
  facilities,
  from,
  to,
  date,
  size,
  timeout,
  database,
  verbose,
  apikey
)
}
\arguments{
\item{countries}{Passed from data function for check}

\item{companies}{Passed from data function for check}

\item{facilities}{Passed from data function for check}

\item{from}{Passed from data function for check}

\item{to}{Passed from data function for check}

\item{date}{Passed from data function for check}

\item{size}{Passed from data function for check}

\item{timeout}{Passed from data function for check}

\item{database}{Passed from data function for check}

\item{verbose}{Passed from data function for check}

\item{apikey}{Passed from data function for check}
}
\description{
check_giedata2input
}
