% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_giedatainput}
\alias{check_giedatainput}
\title{Title}
\usage{
check_giedatainput(
  country,
  company,
  facility,
  from,
  to,
  page,
  date,
  size,
  type,
  verbose,
  apikey
)
}
\arguments{
\item{apikey}{}
}
\value{

}
\description{
Title
}
