% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_gielistinginput}
\alias{check_gielistinginput}
\title{check_gielistinginput}
\usage{
check_gielistinginput(region, country, facilities, database, apikey)
}
\arguments{
\item{region}{Passed from listing function for check}

\item{country}{Passed from listing function for check}

\item{facilities}{Passed from listing function for check}

\item{database}{Passed from listing function for check}

\item{apikey}{Passed from listing function for check}
}
\description{
check_gielistinginput
}
