% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_gieunavinput}
\alias{check_gieunavinput}
\title{check_gieunavinput}
\usage{
check_gieunavinput(
  country,
  start,
  end,
  type,
  end_flag,
  timeout,
  size,
  database,
  apikey
)
}
\arguments{
\item{country}{Passed from data function for check}

\item{start}{Passed from data function for check}

\item{end}{Passed from data function for check}

\item{type}{Passed from data function for check}

\item{end_flag}{Passed from data function for check}

\item{timeout}{Passed from data function for check}

\item{size}{Passed from data function for check}

\item{database}{Passed from data function for check}

\item{apikey}{Passed from data function for check}
}
\description{
check_gieunavinput
}
