% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countryinfo}
\alias{countryinfo}
\title{Mapping of country name and two-character country code used by GIE}
\format{
\subsection{\code{countryinfo}}{

A data frame with 17 rows and 2 columns:
\describe{
\item{name}{Country name}
\item{code}{Two-character country code}
}
}
}
\source{
https://www.gie.eu/
}
\usage{
countryinfo
}
\description{
A dataset to help you switch between country name and country code
in the usage of 'gie', as gie_listing() and gie_load() work
with different modes of the country (name vs. two-character code).
}
\keyword{datasets}
