% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_listelements}
\alias{extract_listelements}
\title{extract_listelements}
\usage{
extract_listelements(listelement)
}
\arguments{
\item{listelement}{List element to extract from}
}
\description{
extract_listelements
}
