% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_giedata.R
\name{get_giedata}
\alias{get_giedata}
\title{Function to download data from GIE's AGSI+ API}
\usage{
get_giedata(
  country = NULL,
  company = NULL,
  facility = NULL,
  from = NULL,
  to = NULL,
  page = 1,
  date = NULL,
  size = 20,
  type = "agsi",
  verbose = FALSE,
  apikey
)
}
\arguments{
\item{country}{Character string. Specify the country of interest.}

\item{company}{Character string. EIC code for the requested company.}

\item{facility}{Character string. EIC code for the requested facility.}

\item{from}{Character string. Specify the start of the time span you are \cr
interested in downloading (format: YYYY-MM-DD).}

\item{apikey}{}
}
\value{
A data frame.
}
\description{
Function to download data from GIE's AGSI+ API
}
