% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_giedatabulk.R
\name{get_giedatabulk}
\alias{get_giedatabulk}
\title{get_giedatabulk}
\usage{
get_giedatabulk(
  countries = NULL,
  companies = NULL,
  facilities = NULL,
  from = NULL,
  to = NULL,
  page = 1,
  date = NULL,
  size = 30,
  type = NULL,
  timeout = 3,
  database = "agsi",
  verbose = FALSE,
  apikey = Sys.getenv("GIE_APIKEY")
)
}
\arguments{
\item{countries}{A character vector of countries in short format (e.g., "DE")
to get data from. Must be of length one (i.e., one country) if you want to
specify multiple companies and/or multiple facilities.}

\item{companies}{A character vector of company EIC codes to get data from. \cr
Must be of length one (i.e., one company) if you want to specify multiple \cr
facilities.}

\item{facilities}{A character vector of facility EIC codes to get data from.}

\item{from}{Character. Specify the start of the time span you are \cr
interested in downloading (format: YYYY-MM-DD).}

\item{to}{Character. Specify the end of the time span you are \cr
interested in downloading (format: YYYY-MM-DD).}

\item{page}{Integer. The page of a multi-page query you want to get.}

\item{date}{Character. If you want to have data only for one date. \cr
If you set 'date', you cannot set the 'from' and/or 'to' parameters \cr
(format: YYYY-MM-DD).}

\item{size}{Integer. The number of results per page.}

\item{type}{Type of facility or company.}

\item{timeout}{Numeric. If the amount of pages of your request exceeds 60, a timeout \cr
will be enforced to prevent the API from timing out. Defaults to 3 seconds, any \cr
values must be set in seconds, too.}

\item{database}{Character. The type of API you want to address ('agsi' or 'alsi').}

\item{verbose}{Logical. Prints information on function progress to the console (default: FALSE).}

\item{apikey}{Character. Your personal API key.}
}
\value{
A data.frame with results
}
\description{
Function to download data from GIE's AGSI+ API in bulk
}
\examples{
\dontrun{
get_giedatabulk(countries = c("DE", "AT", "FR"), date = "2022-04-01")
}

}
