% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gielisting.R
\name{get_gielisting}
\alias{get_gielisting}
\title{get_gielisting}
\usage{
get_gielisting(
  show = "listing",
  region = NULL,
  country = NULL,
  facilities = FALSE,
  type = "agsi",
  apikey
)
}
\arguments{
\item{apikey}{}
}
\value{

}
\description{
Function to download raw or parsed results for the countries, \cr
companies and facilities available from the AGSI+ API of GIE. The EIC codes \cr
of the results can be used to in turn download the actual data \cr
using \code{get_giedata()}.
}
\examples{

}
