% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_listinghierarchy}
\alias{get_listinghierarchy}
\title{get_listinghierarchy}
\usage{
get_listinghierarchy(raw_results, region, country, facilities)
}
\arguments{
\item{facilities}{}
}
\value{

}
\description{
get_listinghierarchy
}
