% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getrequest}
\alias{getrequest}
\title{Function to execute GET request}
\usage{
getrequest(
  country,
  company,
  facility,
  from,
  to,
  page,
  date,
  size,
  type,
  database,
  pages = NULL,
  verbose,
  apikey
)
}
\arguments{
\item{apikey}{}
}
\value{

}
\description{
Function to execute GET request
}
