% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gie_aggregate.R
\name{gie_aggregate}
\alias{gie_aggregate}
\title{Get the aggregated historical data export for Europe or Non Europe}
\usage{
gie_aggregate(area, api_key = NULL)
}
\arguments{
\item{area}{eu for Europe, ne for Non Europe.}

\item{api_key}{The default is NULL and searches for your GIE_PAT in you .Renviron
file.}
}
\description{
Get the aggregated historical data export for Europe or Non Europe
}
\examples{
{

library(gie)

ne <- gie_aggregate("ne")
eu <- gie_aggregate("eu")

}

}
