% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gie_aggregate.R
\name{gie_gas_aggregate}
\alias{gie_gas_aggregate}
\title{Get the aggregated historical data export for Europe or Non Europe}
\usage{
gie_gas_aggregate(area, api_key = NULL, ...)
}
\arguments{
\item{area}{eu for Europe, ne for Non Europe.}

\item{api_key}{The default is NULL and searches for your GIE_PAT in you .Renviron
file.}
}
\description{
Get the aggregated historical data export for Europe or Non Europe
}
\examples{
{

library(gie)
library(tidyverse)

ne <- gie_gas_aggregate("ne")
eu <- gie_gas_aggregate("eu")

eu \%>\%
  mutate(yr = year(gasDayStartedOn),
         mnth = month(gasDayStartedOn),
         mday = mday(gasDayStartedOn)) \%>\%
  filter(mnth == 7) \%>\%
  group_by(yr) \%>\%
  filter(mday == max(mday)) \%>\%
  ungroup() \%>\%
  ggplot(., aes(gasDayStartedOn, gasInStorage)) +
  geom_col() +
  geom_point(aes(gasDayStartedOn, workingGasVolume)) +
  labs(title = "End of June Gas Storage")
}

}
