% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gie_country.R
\name{gie_gas_country}
\alias{gie_gas_country}
\title{Get the aggregated historical data export for for a specific country.}
\usage{
gie_gas_country(country_code, api_key = NULL, ...)
}
\arguments{
\item{country_code}{Two digit country code. Ex: NL, DE, DK, SE, FI etc.}

\item{api_key}{The default is NULL and searches for your GIE_PAT in you .Renviron
file.}
}
\description{
Get the aggregated historical data export for for a specific country.
}
\examples{
{

library(gie)

de <- gie_country("DE")
nl <- gie_country("NL")

}

}
