% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gie_date.R
\name{gie_gas_date}
\alias{gie_gas_date}
\title{Get the table with storage data for a specific date.}
\usage{
gie_gas_date(date = Sys.Date() - 1, api_key = NULL)
}
\arguments{
\item{date}{Iso date.}

\item{api_key}{The default is NULL and searches for your GIE_PAT in you .Renviron
file.}
}
\description{
Get the table with storage data for a specific date.
}
