% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gie_eic.R
\name{gie_gas_eic}
\alias{gie_gas_eic}
\title{Get the historical data export for for a specific company within a country.}
\usage{
gie_gas_eic(country_code, eic_code, api_key = NULL, ...)
}
\arguments{
\item{country_code}{Two digit country code. Ex: NL, DE, DK, SE, FI etc.}

\item{eic_code}{The 21 digit eic code of the company as found on the API page.}

\item{api_key}{The default is NULL and searches for your GIE_PAT in you .Renviron
file.}
}
\description{
Get the historical data export for for a specific company within a country.
}
\examples{

library(tidyverse)
library(gie)

head(gie_gas_eic("AT", "25X-GSALLC-----E"))

}
