% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gie_eic.R
\name{gie_gas_eic_list_full}
\alias{gie_gas_eic_list_full}
\title{Function to get a list of EIC units with some more info.}
\usage{
gie_gas_eic_list_full()
}
\description{
Function to get a list of EIC units with some more info.
}
