% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gie_company_eic.R
\name{gie_gas_facility_eic}
\alias{gie_gas_facility_eic}
\title{Get the historical data export for for a specific facility from a company within a country.}
\usage{
gie_gas_facility_eic(
  country_code,
  company_code,
  facility_code,
  api_key = NULL,
  max_pages = 5000
)
}
\arguments{
\item{country_code}{Two digit country code. Ex: NL, DE, DK, SE, FI etc.}

\item{company_code}{The 21 digit eic code of the company as found on the API page.}

\item{facility_code}{The 21 digit eic code of the facility as found on the API page.}

\item{api_key}{The default is NULL and searches for your GIE_PAT in you .Renviron
file.}

\item{max_pages}{Maximum pages. Defaults to 5000 to get all pages.}
}
\description{
Get the historical data export for for a specific facility from a company within a country.
}
\examples{

library(gie)

tail(gie_gas_facility_eic(country_code = "DE", company_code = "21X000000001160J"
                          facility_code = "21Z000000000271O"))


}
