% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gie_iip.R
\name{gie_iip_recent}
\alias{gie_iip_recent}
\title{GIE IIP JSON Feed}
\usage{
gie_iip_recent()
}
\description{
JSON feed for the Urgent Market Messages sent
    in on the GIE REMIT platform, consisting
    of Type 2 - Unavailabilities of Gas
    Facilities and Type 3 - Other Market
    Information UMMs. Default view is data submitted today.
}
