% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gie_internal_page_request.R
\name{gie_internal_page_request_lng}
\alias{gie_internal_page_request_lng}
\title{Get the historical data export for for a specific facility from a company within a country.}
\usage{
gie_internal_page_request_lng(url, api_key, max_pages = 10, country_code)
}
\arguments{
\item{url}{The base URL to request}

\item{api_key}{The default is NULL and searches for your GIE_PAT in you .Renviron
file.}

\item{max_pages}{Page results if possible, up to this number of pages.
Both a message and a warning are emitted if there are more pages
than this parameter specifies.}
}
\description{
Get the historical data export for for a specific facility from a company within a country.
}
