% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gie_listinghierarchy}
\alias{gie_listinghierarchy}
\title{gie_listinghierarchy}
\usage{
gie_listinghierarchy(raw_results, region, country, facilities, database)
}
\arguments{
\item{raw_results}{Raw results from API call}

\item{region}{Region to filter for}

\item{country}{Country to filter for}

\item{facilities}{Should facilties be exported as well}

\item{database}{Database to get info from}
}
\description{
gie_listinghierarchy
}
