% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gie_company_eic.R
\name{gie_lng_company_eic}
\alias{gie_lng_company_eic}
\title{Get the historical data export for for a specific facility from a company within a country.}
\usage{
gie_lng_company_eic(country_code, eic_code, eic_company_code, api_key = NULL)
}
\arguments{
\item{country_code}{Two digit country code. Ex: NL, DE, DK, SE, FI etc.}

\item{eic_code}{The 21 digit eic code of the facility as found on the API page.}

\item{eic_company_code}{The 21 digit eic code of the company as found on the API page.}

\item{api_key}{The default is NULL and searches for your GIE_PAT in you .Renviron
file.}
}
\description{
Get the historical data export for for a specific facility from a company within a country.
}
\examples{
{

library(gie)

eic <- gie_lng_company_eic(country_code = "DE",
                           eic_code = "21W000000000100J",
                           eic_company_code = "21X000000001368W")

}

}
