% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gie_eic.R
\name{gie_lng_eic_list}
\alias{gie_lng_eic_list}
\title{Function to get a list of EIC units.}
\usage{
gie_lng_eic_list(unnest_facilities = TRUE)
}
\arguments{
\item{unnest_facilities}{Defaults to true.}
}
\description{
Function to get a list of EIC units.
}
