% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gie_internal_page_request.R
\name{gie_pagination_api}
\alias{gie_pagination_api}
\title{Get the data from the API endpoint.}
\usage{
gie_pagination_api(
  hostname,
  continent = NULL,
  country = NULL,
  company = NULL,
  facility = NULL,
  page = NULL,
  max_pages = 5000,
  api_key = NULL
)
}
\arguments{
\item{hostname}{Hostname}

\item{continent}{Continent}

\item{country}{Country}

\item{company}{Company}

\item{facility}{Facility}

\item{page}{Page}

\item{max_pages}{Max pages.}

\item{api_key}{Api key}
}
\description{
Get the data from the API endpoint.
}
