% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{giedata_parseresult}
\alias{giedata_parseresult}
\title{Process results from GET request}
\usage{
giedata_parseresult(raw_results, country)
}
\arguments{
\item{country}{}
}
\value{

}
\description{
Process results from GET request
}
