% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_date}
\alias{is_valid_date}
\title{Check if a string is a valid date}
\usage{
is_valid_date(date_string)
}
\arguments{
\item{date_string}{Date string to check}
}
\value{
TRUE or FALSE
}
\description{
Check if a string is a valid date
}
