% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_unav}
\alias{parse_unav}
\title{parse_unav}
\usage{
parse_unav(raw_results)
}
\arguments{
\item{raw_results}{Raw results}
}
\value{
A data.frame
}
\description{
A function to parse unavailability results
}
