% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{send_getrequest}
\alias{send_getrequest}
\title{send_getrequest}
\usage{
send_getrequest(url, apikey, ...)
}
\arguments{
\item{url}{URL for GET request}

\item{apikey}{API key}

\item{...}{Further API parameters}
}
\description{
send_getrequest
}
