% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_html}
\alias{strip_html}
\title{strip_html}
\usage{
strip_html(string)
}
\arguments{
\item{string}{}
}
\value{
A vector with HTML decoded text
}
\description{
A function to HTML decode a character vector of length > 1
}
