% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_GET}
\alias{try_GET}
\title{try_GET}
\usage{
try_GET(url, apikey, ...)
}
\arguments{
\item{url}{URL for API call}

\item{apikey}{API key}

\item{...}{Further parameter for httr::GET}
}
\description{
try_GET
}
