% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gie_internal_page_request.R
\name{url_api_construct}
\alias{url_api_construct}
\title{Construct the url.}
\usage{
url_api_construct(
  hostname,
  continent = NULL,
  country = NULL,
  company = NULL,
  facility = NULL,
  page = NULL,
  size = 300,
  scheme = "https",
  path = "api"
)
}
\arguments{
\item{hostname}{Hostname}

\item{continent}{Continent}

\item{country}{Country}

\item{company}{Company}

\item{facility}{Facility}

\item{page}{Page}

\item{size}{Size. Defaults to 300.}

\item{scheme}{Scheme. Defaults to 'https'}

\item{path}{Path. Defaults to 'api'}
}
\description{
Construct the url.
}
