% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_giedatainput}
\alias{check_giedatainput}
\title{check_giedatainput}
\usage{
check_giedatainput(
  country,
  company,
  facility,
  from,
  to,
  date,
  size,
  timeout,
  database,
  verbose,
  apikey
)
}
\arguments{
\item{country}{Passed from data function for check}

\item{company}{Passed from data function for check}

\item{facility}{Passed from data function for check}

\item{from}{Passed from data function for check}

\item{to}{Passed from data function for check}

\item{date}{Passed from data function for check}

\item{size}{Passed from data function for check}

\item{timeout}{Passed from data function for check}

\item{database}{Passed from data function for check}

\item{verbose}{Passed from data function for check}

\item{apikey}{Passed from data function for check}
}
\description{
check_giedatainput
}
