% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_giedata.R
\name{get_giedata}
\alias{get_giedata}
\title{Function to download data from GIE's AGSI+ API}
\usage{
get_giedata(
  country = NULL,
  company = NULL,
  facility = NULL,
  from = NULL,
  to = NULL,
  page = 1,
  date = NULL,
  size = 30,
  type = NULL,
  database = "agsi",
  verbose = FALSE,
  apikey
)
}
\arguments{
\item{country}{Character. Specify the country of interest.}

\item{company}{Character. EIC code for the requested company.}

\item{facility}{Character. EIC code for the requested facility.}

\item{from}{Character. Specify the start of the time span you are \cr
interested in downloading (format: YYYY-MM-DD).}

\item{to}{Character. Specify the end of the time span you are \cr
interested in downloading (format: YYYY-MM-DD).}

\item{page}{Integer. The page of a multi-page query you want to get.}

\item{date}{Character. If you want to have data only for one date. \cr
If you set 'date', you cannot set the 'from' and/or 'to' parameters \cr
(format: YYYY-MM-DD).}

\item{size}{Integer. The number of results per page.}

\item{database}{Character. The type of API you want to address ('agsi' or 'alsi').}

\item{apikey}{Character. Your personal API key.}
}
\value{
A data.frame or list with the results.
}
\description{
Function to download data from GIE's AGSI+ API
}
