% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giedata_getgasdata.R
\name{get_giedata}
\alias{get_giedata}
\title{Function to download data from GIE's AGSI+ API}
\usage{
get_giedata(
  country,
  from = Sys.Date() - 5,
  to = Sys.Date() - 1,
  page = 1,
  date = NULL,
  size = 20,
  type = "agsi",
  apikey
)
}
\arguments{
\item{country}{Character string. Specify the country of interest.}

\item{from}{Character string. Specify the start of the time span you are \cr
interested in downloading (format: YYYY-MM-DD).}

\item{apikey}{}
}
\value{
A data frame.
}
\description{
Function to download data from GIE's AGSI+ API
}
