% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_listinghierarchy}
\alias{get_listinghierarchy}
\title{get_listinghierarchy}
\usage{
get_listinghierarchy(raw_results, region, country, facilities)
}
\arguments{
\item{raw_results}{Raw results from API call}

\item{region}{Region to filter for}

\item{country}{Country to filter for}

\item{facilities}{Should facilties be exported as well}
}
\description{
get_listinghierarchy
}
