% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseresult}
\alias{parseresult}
\title{parseresult}
\usage{
parseresult(raw_results, database)
}
\arguments{
\item{raw_results}{Raw results from GET request}

\item{database}{The database to be used}
}
\description{
parseresult
}
